% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{access_to_npc}
\alias{access_to_npc}
\title{Predict the nets per capita associated witha given access}
\usage{
access_to_npc(access, type = "loess")
}
\arguments{
\item{access}{A single or vector of access values}

\item{type}{The npc to access model to use. This may be:
\itemize{
\item{"loess"}{: a loess fit to observed access-npc data}
\item{"loess_extrapolate"}{: a loess fit to observed access-npc data with trends extrapolated above and below
observed values.}
\item{"linear"}{: a linear fit to the observed access-npc data, fitted to the trend for observeation with access < 0.5}
}}
}
\value{
Predicted nets per capita
}
\description{
Predict the nets per capita associated witha given access
}
