% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{access_to_usage}
\alias{access_to_usage}
\title{Convert access to usage}
\usage{
access_to_usage(access, use_rate)
}
\arguments{
\item{access}{A single value or vector of access.}

\item{use_rate}{A single value or vector of usage rates.}
}
\value{
Usage
}
\description{
Convert access to usage
}
