% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{anpcd_to_npc}
\alias{anpcd_to_npc}
\title{Convert annual nets per capital delivered to nets per capita}
\usage{
anpcd_to_npc(
  anpcd,
  distribution_freq,
  half_life,
  net_loss_function = net_loss_map,
  ...
)
}
\arguments{
\item{anpcd}{Annual nets per capital delivered}

\item{distribution_freq}{A single distribution frequency of nets in days. Default = 3*365 (3 years).}

\item{half_life}{Net loss half life (days)}

\item{net_loss_function}{Option to choose between exponential net loss (net_loss_exp) or
the smooth-compact net loss function from Bertozzi-Villa, Amelia, et al.
Nature communications 12.1 (2021): 1-12 (net_loss_map). Default = net_loss_map.}

\item{...}{additional arguments for the net_loss_function}
}
\value{
Nets per capita
}
\description{
Convert annual nets per capital delivered to nets per capita
}
