% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_to_usage_draft.R
\name{convert_annual_nets_distributed_to_npc}
\alias{convert_annual_nets_distributed_to_npc}
\title{Convert annual nets distributed per capita to equilibrium nets per capita ###}
\usage{
convert_annual_nets_distributed_to_npc(
  nets_distributed,
  distribution_freq = 3 * 365,
  half_life_data = prepare_data()$half_life_data,
  net_loss_function = net_loss_exp,
  k = 20
)
}
\arguments{
\item{distribution_freq}{A single distribution frequency of nets in days. Default = 3*365 (3 years).}

\item{half_life_data}{Dataset of country-specific net half lives obtained with prepare_data() or
a value/vector of net half lives (median retention times) in days. Can also
be limited to specific ISO3 codes.}

\item{net_loss_function}{Option to choose between exponential net loss (net_loss_exp) or
the smooth-compact net loss function from Bertozzi-Villa, Amelia, et al.
Nature communications 12.1 (2021): 1-12 (net_loss_map). Default = net_loss_exp.}

\item{k}{Fixed rate parameter of the MAP function of net loss (net_loss_map()), used to
estimate the median net retention times by country. Default = 20 from
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.}

\item{nets_distibuted}{Dataframe of target annual nets dstributed per capita}
}
\description{
Convert annual nets distributed per capita to equilibrium nets per capita ###
}
