% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_to_usage.R
\name{convert_nets_to_usage}
\alias{convert_nets_to_usage}
\title{Convert annual nets distributed per capita to estimated usage}
\usage{
convert_nets_to_usage(
  target_nets,
  distribution_freq = 3 * 365,
  use_rate_data = prepare_data()$use_rate_by_country,
  half_life_data = prepare_data()$half_life_data,
  access_vs_npc_data = prepare_data()$loess_for_prediction,
  extrapolate_npc = "loess",
  net_loss_function = net_loss_map,
  k = 20
)
}
\arguments{
\item{target_nets}{A single value or vector of
desired target annual per capita net distributions to model.}

\item{distribution_freq}{A single distribution frequency of nets in days.
Default = 3*365 (3 years).}

\item{use_rate_data}{Dataset of country-specific net use rates obtained with
prepare_data() or a value/vector of net use rates.
Default = prepare_data()$use_rate_by_country. Can also be limited to specific
ISO3 codes.}

\item{half_life_data}{Dataset of country-specific net half lives obtained
with prepare_data() or a value/vector of net half lives (median retention
times) in days. Can also be limited to specific ISO3 codes.}

\item{access_vs_npc_data}{Dataset of net access against nets per capita
(Loess curve) obtained with prepare_data().
Default = prepare_data()$loess_for_prediction. Needs to include all
countries/ISO3 codes.}

\item{extrapolate_npc}{Option to extrapolate target nets per capita beyond
fitted Loess curve. Default = "loess" for continuation of Loess curve trend.
Use option "linear" for linear extrapolation. For any other inputs, NA
outputs are returned for any target usages exceeding currently observed
levels.}

\item{net_loss_function}{Option to choose between exponential net loss
(net_loss_exp) or the smooth-compact net loss function from Bertozzi-Villa,
Amelia, et al. Nature communications 12.1 (2021): 1-12 (net_loss_map).
Default = net_loss_map.}

\item{k}{Fixed rate parameter of the MAP function of net
loss (net_loss_map()), used to estimate the median net retention times by
country. Default = 20 from Bertozzi-Villa, Amelia, et al. Nature
communications 12.1 (2021): 1-12.}
}
\value{
Dataframe of usage, access, and nets per capita for the desired
input nets distributed by country.
}
\description{
Calculates the estimated net usage from the annual number of nets distributed
per capita. Does the reverse of usage_to_annual_nets_distributed().
}
\details{
Nets distributed is converted to nets per capita using half life data of
nets, which accounts for net loss over time within a distribution cycle. Nets
per capita is then converted to access using the observed relationship
between access and nets per capita across Africa, as described in
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12. Then
access is converted to usage using the input use rate(s).
}
