% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nets_to_usage.R, R/npc_to_usage_draft.R
\name{convert_npc_to_usage}
\alias{convert_npc_to_usage}
\title{Convert nets per capita to bednet usage}
\usage{
convert_npc_to_usage(npc, iso3c = "all", extrapolate_npc = "linear", k = 20)

convert_npc_to_usage(npc, iso3c = "all", extrapolate_npc = "linear", k = 20)

convert_npc_to_usage(npc, iso3c = "all", extrapolate_npc = "linear", k = 20)
}
\arguments{
\item{extrapolate_npc}{Option to extrapolate target nets per capita beyond
fitted Loess curve. Default = "loess" for continuation of Loess curve trend.
Use "FALSE" for no extrapolation. For any other inputs, NA
outputs are returned for any target usages exceeding currently observed
levels.}

\item{k}{Fixed rate parameter of the MAP function of net
loss (net_loss_map()), used to estimate the median net retention times by
country. Default = 20 from Bertozzi-Villa, Amelia, et al. Nature
communications 12.1 (2021): 1-12.}

\item{nets_to_npc_output}{Output from
convert_annual_nets_distributed_to_npc(). Dataframe of target nets
distributed, and estimated mean nets per capita using the net loss function
and distribution frequency by country.}
}
\value{
Dataframe of access, nets per capita and annual nets distributed per capita
for the desired steady-state target usage by country
}
\description{
Calculates the net usage from the estimated nets per capita.
Nets per capita is converted to access using the observed relationship
between access and nets per capita across Africa, as described in
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.
The derived access is then converted to usage using the input use rate(s).

Calculates the annual number of nets needed per capita to achieve a desired
net usage.
Usage is converted to access using the input use rate(s), which is then
converted to nets per capita using the observed relationship between
access and nets per capita across Africa, as described in
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.

Calculates the annual number of nets needed per capita to achieve a desired
net usage.
Usage is converted to access using the input use rate(s), which is then
converted to nets per capita using the observed relationship between
access and nets per capita across Africa, as described in
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.
}
