% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_to_annual_nets_distributed.R
\name{convert_usage_to_npc}
\alias{convert_usage_to_npc}
\title{Convert bednet usage to nets per capita}
\usage{
convert_usage_to_npc(
  target_usage,
  use_rate_data = prepare_data()$use_rate_by_country,
  access_vs_npc_data = prepare_data()$loess_for_prediction,
  extrapolate_npc = "loess"
)
}
\arguments{
\item{target_usage}{A single value or vector of desired target usages to model.}

\item{use_rate_data}{Dataset of country-specific net use rates obtained with prepare_data() or
a value/vector of net use rates. Default = prepare_data()$use_rate_by_country. Can also
be limited to specific ISO3 codes.}

\item{access_vs_npc_data}{Dataset of net access against nets per capita (Loess curve)
obtained with prepare_data(). Default = prepare_data()$loess_for_prediction. Needs to
include all countries/ISO3 codes.}

\item{extrapolate_npc}{Option to extrapolate target nets per capita beyond fitted
Loess curve. Default = "loess" for continuation of Loess curve trend. Use option
"linear" for linear extrapolation. For any other inputs, NA outputs are returned
for any target usages exceeding currently observed levels.}
}
\description{
Calculates the annual number of nets needed per capita to achieve a desired
net usage.
Usage is converted to access using the input use rate(s), which is then
converted to nets per capita using the observed relationship between
access and nets per capita across Africa, as described in
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.
}
