% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_conversion.R
\name{crop_to_distribution_dynamic}
\alias{crop_to_distribution_dynamic}
\title{Dynamic crop to distribution estimation}
\usage{
crop_to_distribution_dynamic(crop, net_loss_function, ...)
}
\arguments{
\item{crop}{Vector fo annual crop estimates}

\item{net_loss_function}{Net loss function}

\item{...}{Additional parameters to the net loss function}
}
\value{
Annual distribution vector
}
\description{
Estimates the distributions required to achieve a given, non-equilibrium
crop.
}
