% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_conversion.R
\name{distribution_to_crop_dynamic}
\alias{distribution_to_crop_dynamic}
\title{Dynamic distribution to crop estimation}
\usage{
distribution_to_crop_dynamic(distribution, net_loss_function, ...)
}
\arguments{
\item{distribution}{Vector of annual distribution estimates}

\item{net_loss_function}{Net loss function}

\item{...}{Additional parameters to the net loss function}
}
\value{
Annual crop vector
}
\description{
Estimates the crop resulting from a given, non-equilibrium
distribution
}
