% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_conversion.R
\name{dynamic_crop}
\alias{dynamic_crop}
\title{Cumulative crop}
\usage{
dynamic_crop(x, xt, timesteps, loss_function, ...)
}
\arguments{
\item{x}{Input distribution}

\item{xt}{Input time steps}

\item{timesteps}{Total timesteps}

\item{loss_function}{Net loss function}

\item{...}{Additional arguments to net loss function}
}
\value{
Sum of multiple, shifted net loss
}
\description{
Cumulative crop
}
