% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_conversion.R
\name{dynamic_opti}
\alias{dynamic_opti}
\title{Optimise distribution}
\usage{
dynamic_opti(
  target,
  target_t,
  xt,
  x_upper = rep(3, length(xt)),
  x_init = rep(0, length(xt)),
  x_lower = rep(0, length(xt)),
  timesteps = max(c(target_t, xt)),
  loss_function,
  ...
)
}
\arguments{
\item{target}{Target values}

\item{target_t}{Target value timesteps}

\item{xt}{Input time steps}

\item{x_upper}{Upper bound on x}

\item{x_init}{Initial values of x}

\item{x_lower}{Lower bound on x}

\item{timesteps}{Total timesteps}

\item{loss_function}{Net loss function}

\item{...}{Additional arguments to net loss function}
}
\value{
Optimised distribution to achieve a given crop
}
\description{
Optimise distribution
}
