% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_usage.R
\name{exp_loss_equilibrium}
\alias{exp_loss_equilibrium}
\title{Estimate average population usage given at equilibrium}
\usage{
exp_loss_equilibrium(distribution, distribution_frequency, mean_retention)
}
\arguments{
\item{distribution}{The \% of the population who are distributed a net}

\item{distribution_frequency}{The frequency of the distribution cycle (days)}

\item{mean_retention}{The average duration of net retention (days)}
}
\value{
The equilibrium population-level net usage
}
\description{
Estimate average population usage given at equilibrium
}
