% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_to_annual_nets_distributed.R
\name{find_usage}
\alias{find_usage}
\title{Convert annual nets distributed per capita to bednet usage}
\usage{
find_usage(
  annual_percapita_nets_distributed,
  country_iso3 = "all",
  extrapolate_npc = "linear",
  k = 20
)
}
\arguments{
\item{annual_percapita_nets_distributed}{A single value or vector of nets distributed per capita to model.}

\item{country_iso3}{Vector of ISO3 codes of African countries for which to return results.
Default = "all".}

\item{extrapolate_npc}{Option to extrapolate target nets per capita beyond fitted
Loess curve. Default = "linear" for linear extrapolation. For any other inputs,
NA outputs are returned for any target usages exceeding currently observed levels.}

\item{k}{Fixed rate parameter of the internal function of net loss, used to
estimate the median net retention times by country. Default = 20 from
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.}
}
\value{
Dataframe of access, nets per capita and net usage for the desired
steady-state annual nets distributed per capita by country.
}
\description{
Assumptions:
\itemize{
\item Net use rates stays at 2019 levels for all countries
\item When converting from NPC to access, use the 2020 loess curve fit of access-NPC
at the country-month level.
}
}
