% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_usage}
\alias{fit_usage}
\title{Fit distribution to match target usage using non-linear optimisation. This
assumes an exponentially distributed net loss function and randomly
correlated net distribution.}
\usage{
fit_usage(
  target_usage,
  target_usage_timesteps,
  distribution_timesteps,
  distribution_init = rep(0, length(distribution_timesteps)),
  distribution_lower = rep(0, length(distribution_timesteps)),
  distribution_upper = rep(1, length(distribution_timesteps)),
  timesteps = max(c(target_usage_timesteps, distribution_timesteps)),
  half_life = 5 * 365,
  ...
)
}
\arguments{
\item{target_usage}{Target usage}

\item{target_usage_timesteps}{Target usage time points}

\item{distribution_timesteps}{A vector of distribution time steps}

\item{distribution_init}{Starting distribution for optimiation}

\item{distribution_lower}{Lower bound on distributions (default = 0)}

\item{distribution_upper}{Upper bound on distribution (default = 1)}

\item{timesteps}{Net loss vector}

\item{half_life}{Net retention half life}

\item{...}{Further arguments to pass to the \link[nloptr]{cobyla}}
}
\value{
Non-linear optimisation fit output
}
\description{
Fit distribution to match target usage using non-linear optimisation. This
assumes an exponentially distributed net loss function and randomly
correlated net distribution.
}
