% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_usage_objective}
\alias{fit_usage_objective}
\title{Internal objective function for least-squares non-linear fitting of target usage}
\usage{
fit_usage_objective(
  distribution,
  distribution_timesteps,
  target_usage,
  target_usage_timesteps,
  mean_retention
)
}
\arguments{
\item{distribution}{A vector of proposed distribution}

\item{distribution_timesteps}{A vector of distribution time steps}

\item{target_usage}{Target usage}

\item{target_usage_timesteps}{Target usage time points}

\item{mean_retention}{The average duration of net retention (days)}
}
\value{
Sum of squared difference between proposed usage and target usage
}
\description{
Internal objective function for least-squares non-linear fitting of target usage
}
