% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_usage_sequential}
\alias{fit_usage_sequential}
\title{Fit distribution to match target usage using a sequential algorithm. This
assumes an exponentially distributed net loss function and randomly
correlated net distribution. This is faster than the nloptr approach,
but does rely on consistent data quality as latter values will be influenced
by earlier values.}
\usage{
fit_usage_sequential(
  target_usage,
  target_usage_timesteps,
  distribution_timesteps,
  distribution_lower = rep(0, length(distribution_timesteps)),
  distribution_upper = rep(1, length(distribution_timesteps)),
  mean_retention = 365 * 5
)
}
\arguments{
\item{target_usage}{Target usage}

\item{target_usage_timesteps}{Target usage time points}

\item{distribution_timesteps}{A vector of distribution time steps}

\item{distribution_lower}{Lower bound on distributions (default = 0)}

\item{distribution_upper}{Upper bound on distribution (default = 1)}

\item{mean_retention}{The average duration of net retention (days)}
}
\description{
Fit distribution to match target usage using a sequential algorithm. This
assumes an exponentially distributed net loss function and randomly
correlated net distribution. This is faster than the nloptr approach,
but does rely on consistent data quality as latter values will be influenced
by earlier values.
}
