% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_usage.R
\name{get_correlation}
\alias{get_correlation}
\title{Get correlation inputs for intervention sampling}
\usage{
get_correlation(population, rho)
}
\arguments{
\item{population}{Population size}

\item{rho}{Correlation parameter}
}
\value{
A names list with the multivariate random normal draw and sigma squared
}
\description{
Code adapted from malariasimulation
}
