% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_halflife}
\alias{get_halflife}
\title{Get half life}
\usage{
get_halflife(iso3c = NA)
}
\arguments{
\item{iso3c}{ISO3c code. If NA will return median half life estimate}
}
\value{
Half life
}
\description{
Get half life
}
