% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_halflife_data}
\alias{get_halflife_data}
\title{Reading in and formatting half life datasets}
\usage{
get_halflife_data(
  path = paste0("https://raw.github.com/", "bertozzivill/map-itn-cube/",
    "publication-2021/paper_figures/", "figure_data/fig_5_llin_half_lives.csv")
)
}
\arguments{
\item{path}{Link to dataset.}
}
\value{
Half life dataset
}
\description{
Reading in and formatting half life datasets
}
