% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_npc_data}
\alias{get_npc_data}
\title{Function for reading in and access and npc data}
\usage{
get_npc_data(
  path = paste0("https://raw.github.com/", "bertozzivill/map-itn-cube/",
    "publication-2021/paper_figures/", "figure_data/fig_4_access_npc.csv")
)
}
\arguments{
\item{path}{Link to dataset.}
}
\value{
Dataset of use rate by country
}
\description{
Function for reading in and access and npc data
}
