% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_usage_rate}
\alias{get_usage_rate}
\title{Get usage rate}
\usage{
get_usage_rate(iso3c = NA)
}
\arguments{
\item{iso3c}{ISO3c code. If NA will return median usage rate estimate}
}
\value{
Usage rate
}
\description{
Get usage rate
}
