% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_usage_rate_data}
\alias{get_usage_rate_data}
\title{Function for reading in and formatting usage rate datasets}
\usage{
get_usage_rate_data(
  path = paste0("https://raw.github.com/", "bertozzivill/map-itn-gts/",
    "master/data/coverage_metrics/", "aggregated_predictions_2019.csv")
)
}
\arguments{
\item{path}{Link to dataset.}
}
\value{
Dataset of use rate by country
}
\description{
Function for reading in and formatting usage rate datasets
}
