% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{halflife}
\alias{halflife}
\title{Retention half life estimates.}
\format{
A data.frame with 480 rows and 11 columns:
}
\usage{
halflife
}
\description{
For original data publication please see:
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.
}
\keyword{datasets}
