% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_to_annual_nets_distributed.R
\name{integrate_net_loss}
\alias{integrate_net_loss}
\title{Integrate net loss function between 0 and 3}
\usage{
integrate_net_loss(k, l)
}
\arguments{
\item{k}{Fixed rate. Default = 20 from Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.}

\item{l}{Country-specific time at which no nets are retained (nets=0).}
}
\description{
Integrate net loss function between 0 and 3
}
