% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_to_annual_nets_distributed.R
\name{integrate_net_loss_vector}
\alias{integrate_net_loss_vector}
\title{Vectorise function for integration of net loss
Allows application on multiple countries/vector of half-lives}
\usage{
integrate_net_loss_vector(k, l)
}
\arguments{
\item{k}{Fixed rate. Default = 20 from Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.}

\item{l}{Country-specific time at which no nets are retained (nets=0).}
}
\description{
Vectorise function for integration of net loss
Allows application on multiple countries/vector of half-lives
}
