% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariasimulation.R
\name{model_distribution_to_usage}
\alias{model_distribution_to_usage}
\title{Estimate the population-level bed net usage at given times for a set of
bed net distributions at given times. This assumes a specified rate of net loss
and that recipients of multiple rounds are random.}
\usage{
model_distribution_to_usage(
  usage_timesteps,
  distribution,
  distribution_timesteps,
  net_loss_function = net_loss_exp,
  ...
)
}
\arguments{
\item{usage_timesteps}{Target usage time points}

\item{distribution}{Vector of model distribution (\% of total population)}

\item{distribution_timesteps}{A vector of distribution time steps}

\item{net_loss_function}{Function to calculate net retention over time.
Should take time as first argument and return proportion retained.}

\item{...}{Additional arguments passed to net_loss_function}
}
\value{
Usage estimates at timepoints
}
\description{
Estimate the population-level bed net usage at given times for a set of
bed net distributions at given times. This assumes a specified rate of net loss
and that recipients of multiple rounds are random.
}
