% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_to_annual_nets_distributed.R
\name{net_loss}
\alias{net_loss}
\title{Function of net loss over time
Used to estimate the median net retention times by country.}
\usage{
net_loss(t, k, l)
}
\arguments{
\item{t}{Single value or vector of timesteps.}

\item{k}{Fixed rate. Default = 20 from Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.}

\item{l}{Country-specific time at which no nets are retained (nets=0).}
}
\value{
The proportion of nets retained over time.
}
\description{
Function of net loss over time
Used to estimate the median net retention times by country.
}
