% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{nets_per_capita_to_access}
\alias{nets_per_capita_to_access}
\title{Predict the access nets associated with a given nets per capita}
\usage{
nets_per_capita_to_access(nets_per_capita, type = "loess")
}
\arguments{
\item{nets_per_capita}{A single or vector of nets per capita}

\item{type}{The npc to access model to use. This may be:
\itemize{
\item{"loess"}{: a loess fit to observed access-npc data}
\item{"loess_extrapolate"}{: a loess fit to observed access-npc data with trends extrapolated above and below
observed values.}
\item{"linear"}{: a linear fit to the observed access-npc data, fitted to the trend for observeation with access < 0.5}
}}
}
\value{
Predicted access
}
\description{
Predict the access nets associated with a given nets per capita
}
