% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npc_fits}
\alias{npc_fits}
\title{Loess and linear model fits to npc access data.}
\format{
A list with three model files:
}
\usage{
npc_fits
}
\description{
For original model publication please see:
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.
Nets per capita are converted to nets distributed annually by accounting for
net loss over time within a given distribution cycle.
}
\keyword{datasets}
