% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{npc_to_anpcd}
\alias{npc_to_anpcd}
\title{Convert nets per capita to annual nets per capital delivered}
\usage{
npc_to_anpcd(
  npc,
  distribution_freq,
  half_life,
  net_loss_function = net_loss_map,
  ...
)
}
\arguments{
\item{npc}{Nets per capita}

\item{distribution_freq}{A single distribution frequency of nets in days. Default = 3*365 (3 years).}

\item{half_life}{Net loss half life (days)}

\item{net_loss_function}{Option to choose between exponential net loss (net_loss_exp) or
the smooth-compact net loss function from Bertozzi-Villa, Amelia, et al.
Nature communications 12.1 (2021): 1-12 (net_loss_map). Default = net_loss_map.}

\item{...}{additional arguments for the net_loss_function}
}
\value{
Annual nets per capita delivered
}
\description{
Convert nets per capita to annual nets per capital delivered
}
