% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_conversion.R
\name{objective}
\alias{objective}
\title{Cumu objetive function}
\usage{
objective(x, xt, target, target_t, timesteps, loss_function, ...)
}
\arguments{
\item{x}{Input distribution}

\item{xt}{Input time steps}

\item{target}{Target values}

\item{target_t}{Target value timesteps}

\item{timesteps}{Total timesteps}

\item{loss_function}{Net loss function}

\item{...}{Additional arguments to net loss function}
}
\description{
Cumu objetive function
}
