% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_usage.R
\name{population_usage}
\alias{population_usage}
\title{Estimate the time series of population-level bed net usage from a vector of
bed net distributions. This is fixed for an exponentially distributed net
retention, as implemented in malariasimulation}
\usage{
population_usage(distribution, distribution_timesteps, timesteps, half_life)
}
\arguments{
\item{distribution}{A vector of distribution \%s}

\item{distribution_timesteps}{A vector of time steps for distributions}

\item{timesteps}{Timesteps}

\item{half_life}{The half life of the exponentially decaying net loss function (days)}
}
\value{
Usage time series vector
}
\description{
Estimate the time series of population-level bed net usage from a vector of
bed net distributions. This is fixed for an exponentially distributed net
retention, as implemented in malariasimulation
}
