% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Function for reading in and formatting required datasets}
\usage{
prepare_data(
 
    aggregated_annual_data_path = "https://raw.github.com/bertozzivill/map-itn-gts/master/data/coverage_metrics/aggregated_predictions_2019.csv",
  aggregated_annual_data_year = 2019,
 
    access_npc_monthly_data_path = "https://raw.github.com/bertozzivill/map-itn-cube/publication-2021/paper_figures/figure_data/fig_4_access_npc.csv",
 
    half_life_data_path = "https://raw.github.com/bertozzivill/map-itn-cube/publication-2021/paper_figures/figure_data/fig_5_llin_half_lives.csv"
)
}
\arguments{
\item{aggregated_annual_data_path}{Link to dataset containing annual net use rate by country and year.
Must have columns with the following names: iso3, year, month, mean, variable.}

\item{aggregated_annual_data_year}{Single year for which to select net use rate. Must be contained in
the dataset in aggregated_annual_data_path.}

\item{access_npc_monthly_data_path}{Link to dataset containing net access and nets per capita by country, year and month.
Must have columns with the following names: iso3, year, month, access_mean, percapita_nets_mean.}

\item{half_life_data_path}{Link to dataset containing net half lives (median retention times) in years by country.
Must have columns with the following names: iso3, half_life.}
}
\value{
3 datasets of use rate by country, Loess curve of access against nets
per capita, and estimated median net retention times by country in days
}
\description{
Function for reading in and formatting required datasets
}
