% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_usage.R
\name{sample_intervention}
\alias{sample_intervention}
\title{Sample those to receive an intervention with correlation}
\usage{
sample_intervention(p, mvnorm, sigma_squared)
}
\arguments{
\item{p}{Probability of receiving a bed net}

\item{mvnorm}{A vector of multivariate normal (correlated) draws.}

\item{sigma_squared}{Sigma squared}
}
\value{
A boolean vector denoting which individuals will receive the intervention
}
\description{
Sample those to receive an intervention with correlation
}
