% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_conversion.R
\name{shifted}
\alias{shifted}
\title{Shift the net loss function}
\usage{
shifted(dt, l, t)
}
\arguments{
\item{dt}{start time}

\item{l}{net loss vector}

\item{t}{timesteps}
}
\value{
shifted, trimmed net loss
}
\description{
Shift the net loss function
}
