% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{usage_to_access}
\alias{usage_to_access}
\title{Convert usage to access}
\usage{
usage_to_access(usage, use_rate)
}
\arguments{
\item{usage}{A single value or vector of desired target usages to model.}

\item{use_rate}{A single value or vector of usage rates.}
}
\value{
Access
}
\description{
Convert usage to access
}
