% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariasimulation.R
\name{usage_to_model_distribution}
\alias{usage_to_model_distribution}
\title{Estimate the malaraisimulation input distribution from usage
assumes an exponentially distributed net loss function and randomly
correlated net distribution.}
\usage{
usage_to_model_distribution(
  usage,
  usage_timesteps,
  distribution_timesteps,
  distribution_lower = rep(0, length(distribution_timesteps)),
  distribution_upper = rep(1, length(distribution_timesteps)),
  mean_retention = 365 * 5
)
}
\arguments{
\item{usage}{Target usage}

\item{usage_timesteps}{Target usage time points}

\item{distribution_timesteps}{A vector of distribution time steps}

\item{distribution_lower}{Lower bound on distributions (default = 0)}

\item{distribution_upper}{Upper bound on distribution (default = 1)}

\item{mean_retention}{The average duration of net retention (days)}
}
\description{
Estimate the malaraisimulation input distribution from usage
assumes an exponentially distributed net loss function and randomly
correlated net distribution.
}
