% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_to_usage.R
\name{convert_npc_to_usage}
\alias{convert_npc_to_usage}
\title{Convert nets per capita to bednet usage}
\usage{
convert_npc_to_usage(
  nets_to_npc_output,
  use_rate_data = prepare_data()$use_rate_by_country,
  access_vs_npc_data = prepare_data()$loess_for_prediction,
  extrapolate_npc = "loess"
)
}
\arguments{
\item{nets_to_npc_output}{Output from
convert_annual_nets_distributed_to_npc(). Dataframe of target nets
distributed, and estimated mean nets per capita using the net loss function
and distribution frequency by country.}

\item{use_rate_data}{Dataset of country-specific net use rates obtained with
prepare_data() or a value/vector of net use rates.
Default = prepare_data()$use_rate_by_country. Can also be limited to specific
ISO3 codes.}

\item{access_vs_npc_data}{Dataset of net access against nets per capita
(Loess curve) obtained with prepare_data().
Default = prepare_data()$loess_for_prediction. Needs to include all
countries/ISO3 codes.}

\item{extrapolate_npc}{Option to extrapolate target nets per capita beyond
fitted Loess curve. Default = "loess" for continuation of Loess curve trend.
Use option "linear" for linear extrapolation. For any other inputs, NA
outputs are returned for any target usages exceeding currently observed
levels.}
}
\value{
Dataframe of access, nets per capita and annual nets distributed per capita
for the desired steady-state target usage by country
}
\description{
Calculates the net usage from the estimated nets per capita.
Nets per capita is converted to access using the observed relationship
between access and nets per capita across Africa, as described in
Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12.
The derived access is then converted to usage using the input use rate(s).
}
