% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{crop_to_distribution}
\alias{crop_to_distribution}
\title{Convert nets per capita to annual nets per capital delivered}
\usage{
crop_to_distribution(
  crop,
  crop_timesteps,
  distribution_timesteps,
  net_loss_function = netz::net_loss_map,
  min_distribution = NULL,
  max_distribution = NULL,
  ...
)
}
\arguments{
\item{crop}{Vector of nets per capita}

\item{crop_timesteps}{Timesteps of crop estimates}

\item{distribution_timesteps}{Timesteps of distributions}

\item{net_loss_function}{Option to choose between exponential net loss (net_loss_exp) or
the smooth-compact net loss function from Bertozzi-Villa, Amelia, et al.
Nature communications 12.1 (2021): 1-12 (net_loss_map). Default = net_loss_map.}

\item{...}{additional arguments for the net_loss_function}
}
\value{
Annual nets per capita delivered
}
\description{
Convert nets per capita to annual nets per capital delivered
}
