% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{distribution_to_crop}
\alias{distribution_to_crop}
\title{Convert annual nets per capital delivered to nets per capita}
\usage{
distribution_to_crop(
  distribution,
  distribution_timesteps,
  crop_timesteps,
  net_loss_function = netz::net_loss_map,
  ...
)
}
\arguments{
\item{distribution}{Nets per capital delivered}

\item{distribution_timesteps}{Timesteps of distributions}

\item{crop_timesteps}{Timesteps to return crop estimates}

\item{net_loss_function}{Option to choose between exponential net loss (net_loss_exp) or
the smooth-compact net loss function from Bertozzi-Villa, Amelia, et al.
Nature communications 12.1 (2021): 1-12 (net_loss_map). Default = net_loss_map.}

\item{...}{additional arguments for the net_loss_function}
}
\value{
Nets per capita
}
\description{
Convert annual nets per capital delivered to nets per capita
}
