% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_usage}
\alias{fit_usage}
\title{Fit distribution to match target usage using non-linear optimisation}
\usage{
fit_usage(
  target_usage,
  target_usage_timesteps,
  distribution_timesteps,
  distribution_init = rep(0.1, length(distribution_timesteps)),
  distribution_lower = rep(0, length(distribution_timesteps)),
  distribution_upper = rep(1, length(distribution_timesteps)),
  population = 500,
  timesteps = 1:max(c(target_usage_timesteps, distribution_timesteps)),
  net_loss = net_loss_exp(1:timesteps, half_life = 5 * 365),
  rho = 0,
  seed = 1234,
  ...
)
}
\arguments{
\item{target_usage}{Target usage}

\item{target_usage_timesteps}{Target usage time points}

\item{distribution_timesteps}{A vector of distribution time steps}

\item{distribution_lower}{Lower bound on distributions (default = 0)}

\item{distribution_upper}{Upper bound on distribution (default = 1)}

\item{population}{Population}

\item{timesteps}{Timesteps}

\item{net_loss}{Net loss vector}

\item{rho}{betwen round correlation parameters}

\item{seed}{Seed - the output of population_usage is stochastic.}

\item{...}{Further arguments to pass to the \link[nloptr]{cobyla}}
}
\value{
Non-linear optimisation fit output
}
\description{
Fit distribution to match target usage using non-linear optimisation
}
