% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_usage_rate}
\alias{get_usage_rate}
\title{Get usage rate}
\usage{
get_usage_rate(iso3c, not_found_fn = stats::median, ...)
}
\arguments{
\item{iso3c}{ISO3c code.}

\item{not_found_fn}{A function to summarise across all
available half-life estimates to produce an estimate
for countries not in the dataset}

\item{...}{Further arguments passed to not_found_fn}
}
\value{
Usage rate
}
\description{
Get usage rate
}
