% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariasimulation.R
\name{model_distribution_to_usage}
\alias{model_distribution_to_usage}
\title{Estimate the population-level bed net usage at given times for a set of
bed net distributions at given times. This assumes a constant rate of net loss
(as in malariasimulation) and that recipients of multiple rounds are random.}
\usage{
model_distribution_to_usage(
  usage_timesteps,
  distribution,
  distribution_timesteps,
  mean_retention = 365 * 5
)
}
\arguments{
\item{distribution}{A vector of distribution, must be between 0 and 1}

\item{distribution_timesteps}{A vector of time steps for distributions (days)}

\item{mean_retention}{The average duration of net retention (days)}

\item{timesteps}{The half life of the exponentially decaying net loss function (days)}
}
\value{
Usage estimates at timepoints
}
\description{
Estimate the population-level bed net usage at given times for a set of
bed net distributions at given times. This assumes a constant rate of net loss
(as in malariasimulation) and that recipients of multiple rounds are random.
}
