% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_loss.R
\name{net_loss_exp}
\alias{net_loss_exp}
\title{Function of net loss over time from malariasimulation
Net loss is exponential.}
\usage{
net_loss_exp(t, half_life)
}
\arguments{
\item{t}{Single value or vector of timesteps in days.}

\item{half_life}{(Country-specific) half-life of nets in days.}
}
\value{
The proportion of nets retained over time.
}
\description{
Function of net loss over time from malariasimulation
Net loss is exponential.
}
