% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_usage.R
\name{population_usage}
\alias{population_usage}
\title{Estimate the time series of population-level bed net usage from a vector of
bed net distributions.}
\usage{
population_usage(
  net_loss,
  population,
  timesteps,
  distribution,
  distribution_timesteps,
  rho = 0
)
}
\arguments{
\item{net_loss}{A vector of the probability of still having a net over time}

\item{population}{Population size}

\item{timesteps}{Timesteps}

\item{distribution}{A vector of distribution \%s}

\item{distribution_timesteps}{A vector of time steps for distributions}

\item{rho}{The between-round correlation parameter}
}
\value{
Usage time series vector
}
\description{
Note: This is stochastic output
}
