% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Function for reading in and formatting required datasets}
\usage{
prepare_data()
}
\value{
3 datasets of use rate by country, Loess curve of access against nets
per capita, and estimated median net retention times by country in days
}
\description{
Function for reading in and formatting required datasets
}
